add production webpack config
