+++ TODO
@@ -1,0 +1,2 @@
+* set callback for picker in App.js........ or create a new one for each view......?????
+* optimize webpack bundle
